DEV: Move the composer service modification to pre-initializer
To avoid the issue:

```
[PLUGIN discourse-shared-edits] \"service:composer\" has already been initialized and registered as a singleton. Move the modifyClass call earlier in the boot process for changes to take effect.
```
