+++ assets/javascripts/initializers/shared-edits-init.js
@@ -233,63 +233,6 @@ function initWithApi(api) {
       this._super(...arguments);
     },
   });
-
-  api.modifyClass("service:composer", {
-    pluginId: PLUGIN_ID,
-
-    open(opts) {
-      const openResponse = this._super(opts);
-      if (openResponse && openResponse.then) {
-        return openResponse.then(() => {
-          if (opts.action === SHARED_EDIT_ACTION) {
-            setupSharedEdit(this.model);
-          }
-        });
-      }
-    },
-
-    collapse() {
-      if (this.get("model.action") === SHARED_EDIT_ACTION) {
-        return this.close();
-      }
-      return this._super();
-    },
-
-    close() {
-      if (this.get("model.action") === SHARED_EDIT_ACTION) {
-        teardownSharedEdit(this.model);
-      }
-      return this._super();
-    },
-
-    save() {
-      if (this.get("model.action") === SHARED_EDIT_ACTION) {
-        return this.close();
-      }
-      return this._super.apply(this, arguments);
-    },
-
-    @on("init")
-    _listenForClose() {
-      this.appEvents.on("composer:close", () => {
-        this.close();
-      });
-    },
-
-    @observes("model.reply")
-    _handleSharedEdit() {
-      if (this.get("model.action") === SHARED_EDIT_ACTION) {
-        performSharedEdit(this.model);
-      }
-    },
-
-    _saveDraft() {
-      if (this.get("model.action") === SHARED_EDIT_ACTION) {
-        return;
-      }
-      return this._super();
-    },
-  });
 }
 
 export default {
